requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
alembic==1.13.2
arch==7.0.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
contourpy==1.2.1
coreforecast==0.0.12
cycler==0.12.1
exchange_calendars==4.5.5
filelock==3.15.4
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
grpcio==1.65.4
html5lib==1.1
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lightning-utilities==0.11.6
lxml==5.3.0
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
networkx==3.2.1
neuralforecast==1.7.4
numpy==1.26.4
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pandas_market_calendars==4.4.1
patsy==0.5.6
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
protobuf==4.25.4
pyarrow==17.0.0
pyluach==2.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
ray==2.34.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.31
statsmodels==0.14.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
toolz==0.12.1
torch==2.2.2
torchmetrics==1.4.0.post0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
utilsforecast==0.2.2
webencodings==0.5.1
Werkzeug==3.0.3
yarl==1.9.4
yfinance==0.2.41
zipp==3.19.2
